requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
tronpy==0.5.0
python-dotenv==1.1.0
pydantic==2.11.3
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
